changed report and errors tests
